Rename buffered module
